Working Server and Client Code
This is essentially the final product of what I have. The following
program runs with a computer behaving as the server, constantly
searching for a client computer that will connect to it.
In order for the client computer to connect to the server, one must
have the server’s local IP on hand, since it is required to enter that
in when opening ClientTest.java file.

The program is executable by running both ServerTest.java and
ClientTest.java.